third_party/blink/web_tests/external/wpt/mediacapture-extensions/GUM-faceFraming.https-expected.txt

This is a testharness.js-based test.
[PASS] Test getSupportedConstraints().faceFraming
[PRECONDITION_FAILED] Test getCapabilities().faceFraming
  Optional property faceFraming not in capabilities
[PRECONDITION_FAILED] Test getSettings().faceFraming
  Optional property faceFraming not in capabilities
[PASS] Test applyConstraints() with faceFraming
Harness: the test ran to completion.
